Make sure file fit partition before flashing
Partition size is calculated as block count times a block size of 512
for DeviceType 2 and 4096 for DeviceType 8.

Flashing stock android on some devices fail due to some partitions
being too small for the corresponding file.  This is the case on
klimtlte, where sboot.bin does not fit into the BOOTLOADER
partition.

The BOOTLOADER partition on klimtlte has the following properties:

  --- Entry #0 ---
  Binary Type: 0 (AP)
  Device Type: 2 (MMC)
  Identifier: 80
  Attributes: 2 (STL Read-Only)
  Update Attributes: 1 (FOTA)
  Partition Block Size/Offset: 0
  Partition Block Count: 2046
  Partition Name: BOOTLOADER
  Flash Filename: sboot.bin

while the BOOT partition has:

  --- Entry #11 ---
  Binary Type: 0 (AP)
  Device Type: 2 (MMC)
  Identifier: 9
  Attributes: 5 (Read/Write)
  Update Attributes: 1 (FOTA)
  Partition Block Size/Offset: 114688
  Partition Block Count: 16384
  Partition Name: BOOT
  Flash Filename: boot.img

The Attributes entry is the only that differs apart from size and
names.  Maybe the file and partition size check should only be done
for partitions with Attributes: 5 (Read/Write).

Tests done:

* Flash stock android (A500FXXS1CSB2) to a5lte, works fine
* Flash stock android (G930FXXU8ETI2) to herolte (with block
  size 4096), works fine
* Flash stock android (T705XXU1CPL1) to klimtlte, fails due to
  BOOTLOADER partition being smaller than sboot.bin, the file has a
  size of 1148160, and the partition 2046*512=1047552