heimdall: add new --wait arg
If used, then Heimdall will wait until a compatible device is found,
rather than exiting instantly with an error.

Note that this is default behaviour of fastboot, so this change should
make life a little bit easier for people writing scripts that support
flashing to different types of devices.  Exiting instantly is still
default behaviour though, for backward compatibility.