Drop link to Glass Echidna donate page
Benjamin created this great piece of software, and is by far the
largest contributor to the code base.  However, it has been a few
years since he was active, I think it is fair to remove the donate
link by now, at least from this fork that I (try to) maintain.

The copyright information, and the Glass Echidna homepage address,
will stay forever in the various files.